+++ lib/models/global/message_part.dart
@@ -23,12 +23,8 @@ class MessagePart {
   late final String fakeSubject = faker.lorem.words(subject?.split(" ").length ?? 0).join(" ");
   String? get displaySubject {
     if (subject == null) return null;
-    if (ss.settings.redactedMode.value) {
-      if (ss.settings.generateFakeMessageContent.value) {
-        return fakeSubject;
-      } else if (ss.settings.hideContactInfo.value) {
-        return "";
-      }
+    if (ss.settings.redactedMode.value && ss.settings.hideMessageContent.value) {
+      return fakeSubject;
     }
     return text;
   }
@@ -36,12 +32,8 @@ class MessagePart {
   late final String fakeText = faker.lorem.words(text?.split(" ").length ?? 0).join(" ");
   String? get displayText {
     if (text == null) return null;
-    if (ss.settings.redactedMode.value) {
-      if (ss.settings.generateFakeMessageContent.value) {
-        return fakeText;
-      } else if (ss.settings.hideContactInfo.value) {
-        return "";
-      }
+    if (ss.settings.redactedMode.value && ss.settings.hideMessageContent.value) {
+      return fakeText;
     }
     return text;
   }